Moved split() functions into a separate header
