+++ src/Engine/MapInfo.cpp
@@ -1,6 +1,7 @@
 #include "MapInfo.h"
 
 #include <cstdlib>
+#include <cstring>
 #include <sstream>
 #include <string>
