+++ src/Engine/Tables/AutonoteTable.cpp
@@ -1,5 +1,6 @@
 #include "AutonoteTable.h"
 
+#include <cstring>
 #include <string>
 
 #include "Utility/Memory/Blob.h"
