+++ src/Engine/Tables/AwardTable.cpp
@@ -1,5 +1,6 @@
 #include "AwardTable.h"
 
+#include <cstring>
 #include <string>
 
 #include "Utility/Memory/Blob.h"
