+++ src/Engine/Tables/MessageScrollTable.cpp
@@ -1,5 +1,6 @@
 #include "MessageScrollTable.h"
 
+#include <cstring>
 #include <string>
 
 #include "Utility/Memory/Blob.h"
