+++ src/Engine/Tables/QuestTable.cpp
@@ -1,5 +1,6 @@
 #include "QuestTable.h"
 
+#include <cstring>
 #include <string>
 
 #include "Utility/Memory/Blob.h"
