+++ src/Engine/Tables/StorylineTextTable.cpp
@@ -1,9 +1,11 @@
 #include "StorylineTextTable.h"
 
+#include <cstring>
 #include <string>
 
 #include "Utility/Memory/Blob.h"
 #include "Utility/String/Transformations.h"
+#include "Utility/String/Split.h"
 
 StorylineText *pStorylineText;
