+++ src/Engine/Tables/TransitionTable.cpp
@@ -1,5 +1,6 @@
 #include "TransitionTable.h"
 
+#include <cstring>
 #include <string>
 
 #include "Utility/Memory/Blob.h"
