+++ src/Utility/String/Transformations.h
@@ -1,11 +1,7 @@
 #pragma once
 
-#include <cstring>
-#include <algorithm> // For std::find.
 #include <string>
-#include <vector>
-
-std::vector<char*> tokenize(char *input, const char separator);
+#include <string_view>
 
 /**
  * @param str                           String to unquote.
@@ -54,42 +50,6 @@ std::string toHexDump(std::string_view s, size_t groupSize = 0);
 std::string replaceAll(std::string_view text, std::string_view what, std::string_view replacement);
 
 std::string replaceAll(std::string_view text, char what, char replacement);
-
-/**
- * Splits the provided string `s` using separator `sep`, passing `std::string_view` chunks into `consumer`.
- *
- * This function doesn't discard empty chunks, so at least one chunk will always be passed to `consumer` - splitting an
- * empty string produces a single empty chunk.
- *
- * @param s                             String to split.
- * @param sep                           Separator character.
- * @param consumer                      Lambda to pass `std::string_view` chunks into.
- */
-template<class Consumer>
-void split(std::string_view s, char sep, Consumer &&consumer) {
-    const char *pos = s.data();
-    const char *end = s.data() + s.size();
-    while (pos != end + 1) {
-        const char *next = std::find(pos, end, sep);
-
-        consumer(std::string_view(pos, next));
-        pos = next + 1;
-    }
-}
-
-void split(std::string_view s, char sep, std::vector<std::string_view> *result);
-
-inline std::vector<std::string_view> split(std::string_view s, char sep) {
-    std::vector<std::string_view> result;
-    split(s, sep, &result);
-    return result;
-}
-
-inline std::vector<std::string_view> split(const char *s, char sep) {
-    return split(std::string_view(s), sep);
-}
-
-std::vector<std::string_view> split(std::string &&s, char sep) = delete; // Don't dangle!
 
 namespace detail {
